+++ Entity Layer/CourseHistory.cs
@@ -11,11 +11,11 @@ namespace Entity_Layer
         public long Id { get; set; }
         public string Code { get; set; }
         public Department Department { get; set; }
-        public int DepartmentId { get; set; }
+        public long DepartmentId { get; set; }
         public Semister Semister { get; set; }
-        public byte SemisterId { get; set; }
+        public long SemisterId { get; set; }
         public Teacher? Teacher { get; set; }
-        public int? TeacherId { get; set; }
-        public int NthHistory { get; set; }
+        public long? TeacherId { get; set; }
+        public long NthHistory { get; set; }
     }
 }
